backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
langchain==0.2.6
langchain-core==0.2.10
langchain-groq==0.1.4
